+++ src/Runtime/FileController/JsonFileController.cs
@@ -18,7 +18,7 @@ namespace CGame
         private static bool CheckFilePath(string path)
         {
             path = Path.ChangeExtension(path, Extension);
-            return !path.GetPathState().ContainsAll(PathState.Invalid, PathState.Directory);
+            return !path.GetPathState().ContainsAll(PathState.Invalid | PathState.Directory);
         }
 
         public static IEnumerable<T> GetValue<T>(string path)